#240 - Changing formatYear function to utils
