+++ src/services/functions/handlers/index.js
@@ -83,16 +83,3 @@ export function isDateInPeriod(date, period, month, year, tool) {
       return false;
   }
 }
-
-export function formatYear(searchedText) {
-  const date = new Date(searchedText);
-  const year = date.getFullYear().toString().substr(-2);
-  const month = date.getMonth() + 1;
-  let day = date.getDate();
-
-  if (day.toString().length === 1) {
-    day = `0${day}`;
-  }
-
-  return `${day}/${month}/${year}`;
-}
